Probably final version of the project
Fixes and touches all over the place. the hasFrom name was not adjusted according to code review but it is considered in eventual update of the app. A No internet connection checkup is done. An attempt to deal with errors from TrafficAPI but not as through handling of petentional errors at WeatherAPI